Adding <GPU Device_ID> to hw profiling scripts

+++ accelwattch_hw_profiler/profile_validation_perf.sh
@@ -35,13 +35,20 @@ if [ "$GPGPUSIM_SETUP_ENVIRONMENT_WAS_RUN" != "1" ]; then
     exit
 fi
 
+if [ "${1}" == "" ]; then
+    echo "No GPU device ID specified, using default GPU device ID 0. For example: ./profile_validation_power.sh volta 1"
+    exit 1
+fi
+DEVID=${1}
+
 runs_dir="$ACCELSIM_ROOT/../hw_run/device-0/${CUDA_VERSION}"
 if [ -d ${runs_dir} ] ; then
 	rm -r ${runs_dir}
 fi
-
-$ACCELSIM_ROOT/../util/hw_stats/run_hw.py -B rodinia-3.1_validation_hw,parboil_validation,cuda_samples_11.0_validation,cutlass_5_trace_validation,cudaTensorCoreGemm_validation --collect other_stats --nsight_profiler --disable_nvprof
-
+DEFAULT_DEV_ID=`echo $CUDA_VISIBLE_DEVICES`
+export CUDA_VISIBLE_DEVICES=$DEVID
+$ACCELSIM_ROOT/../util/hw_stats/run_hw.py -D $DEVID -B rodinia-3.1_validation_hw,parboil_validation,cuda_samples_11.0_validation,cutlass_5_trace_validation,cudaTensorCoreGemm_validation --collect other_stats --nsight_profiler --disable_nvprof
+export CUDA_VISIBLE_DEVICES=$DEFAULT_DEV_ID
 $ACCELSIM_ROOT/../accelwattch_hw_profiler/gen_hw_perf_csv.py -d ${runs_dir}
 
 mv ${root_dir}/hw_perf.csv $ACCELSIM_ROOT/../accelwattch_hw_profiler/hw_perf.csv
